+++ CHANGELOG.md
@@ -1,6 +1,14 @@
 # Unreleased changes
 
-[Full changelog](https://github.com/mozilla/glean.js/compare/v5.0.0...main)
+[Full changelog](https://github.com/mozilla/glean.js/compare/v5.0.1...main)
+
+# v5.0.1 (2024-04-30)
+
+[Full changelog](https://github.com/mozilla/glean.js/compare/v5.0.0...v5.0.1)
+
+* [#1923](https://github.com/mozilla/glean.js/pull/1923): Bumped `glean_parser` version to `14.0.1`.
+* [#1921](https://github.com/mozilla/glean.js/pull/1921): BUGFIX: Fix issue causing `glean.client.annotation.experimentation_id` metric to not get added in certain pings.
+* [#1919](https://github.com/mozilla/glean.js/pull/1919): Add `glean.page_id` to Glean automatic events.
 
 # v5.0.0 (2024-03-25)